+++ coxeter/shapes/circle.py
@@ -189,3 +189,13 @@ class Circle(Shape2D):
         This is 1 by definition for circles.
         """
         return 1
+
+    @property
+    def minimal_centered_bounding_circle(self):
+        """:class:`~.Circle`: Get the smallest bounding concentric circle."""
+        return Circle(self.radius, self.center)
+
+    @property
+    def minimal_bounding_circle(self):
+        """:class:`~.Circle`: Get the smallest bounding circle."""
+        return Circle(self.radius, self.center)