+++ coxeter/shapes/convex_spheropolyhedron.py
@@ -4,10 +4,13 @@ A convex spheropolyhedron is defined by the Minkowski sum of a convex
 polyhedron and a sphere of some radius.
 """
 
+import warnings
+
 import numpy as np
 
 from .base_classes import Shape3D
 from .convex_polyhedron import ConvexPolyhedron
+from .sphere import Sphere
 
 
 class ConvexSpheropolyhedron(Shape3D):
@@ -31,7 +34,7 @@ class ConvexSpheropolyhedron(Shape3D):
         ...   radius=0.5)
         >>> spherocube.center
         array([0., 0., 0.])
-        >>> sphere = spherocube.circumsphere_from_center
+        >>> sphere = spherocube.minimal_centered_bounding_sphere
         >>> sphere.radius
         2.2320...
         >>> spherocube.gsd_shape_spec
@@ -319,9 +322,12 @@ class ConvexSpheropolyhedron(Shape3D):
         shape distinguishes this sphere from most typical circumsphere
         calculations.
         """  # noqa: E501
-        circumsphere = self.polyhedron.circumsphere_from_center
-        circumsphere.radius += self.radius
-        return circumsphere
+        warnings.warn(
+            "The circumsphere_from_center property is deprecated, use "
+            "minimal_centered_bounding_sphere instead",
+            DeprecationWarning,
+        )
+        return self.minimal_centered_bounding_sphere
 
     @property
     def insphere_from_center(self):
@@ -344,3 +350,15 @@ class ConvexSpheropolyhedron(Shape3D):
         insphere = self.polyhedron.insphere_from_center
         insphere.radius += self.radius
         return insphere
+
+    @property
+    def minimal_bounding_sphere(self):
+        """:class:`~.Sphere`: Get the minimal bounding sphere."""
+        polyhedron_sphere = self.polyhedron.minimal_bounding_sphere
+        return Sphere(polyhedron_sphere.radius + self.radius, polyhedron_sphere.center)
+
+    @property
+    def minimal_centered_bounding_sphere(self):
+        """:class:`~.Sphere`: Get the minimal concentric bounding sphere."""
+        polyhedron_sphere = self.polyhedron.minimal_centered_bounding_sphere
+        return Sphere(polyhedron_sphere.radius + self.radius, polyhedron_sphere.center)