+++ tests/test_ellipse.py
@@ -5,7 +5,8 @@ from hypothesis.extra.numpy import arrays
 from hypothesis.strategies import floats
 from pytest import approx
 
-from coxeter.shapes.ellipse import Ellipse
+from conftest import _test_get_set_minimal_bounding_sphere_radius, sphere_isclose
+from coxeter.shapes import Circle, Ellipse
 
 
 @given(floats(0.1, 1000), floats(0.1, 1000))
@@ -124,6 +125,8 @@ def test_moment_inertia(a, b, center):
     ellipse = Ellipse(a, b)
     assert np.all(np.asarray(ellipse.planar_moments_inertia) >= 0)
 
+    # We must set the center after construction so that the inertia tensor
+    # calculation is not shifted away from the origin.
     ellipse.center = center
     area = ellipse.area
     expected = [np.pi / 4 * a * b ** 3, np.pi / 4 * a ** 3 * b, 0]
@@ -143,6 +146,46 @@ def test_center():
     center = (1, 1, 1)
     ellipse.center = center
     assert all(ellipse.center == center)
+
+
+@given(
+    floats(0.1, 1000),
+    floats(0.1, 1000),
+    arrays(np.float64, (3,), elements=floats(-10, 10, width=64), unique=True),
+)
+def test_minimal_bounding_circle(a, b, center):
+    ellipse = Ellipse(a, b, center)
+    bounding_circle = ellipse.minimal_bounding_circle
+    assert sphere_isclose(bounding_circle, Circle(max(a, b), center))
+
+
+@given(
+    floats(0.1, 1000),
+    floats(0.1, 1000),
+    arrays(np.float64, (3,), elements=floats(-10, 10, width=64), unique=True),
+)
+def test_minimal_centered_bounding_circle(a, b, center):
+    ellipse = Ellipse(a, b, center)
+    bounding_circle = ellipse.minimal_centered_bounding_circle
+    assert sphere_isclose(bounding_circle, Circle(max(a, b), center))
+
+
+@given(
+    floats(0.1, 1000),
+    floats(0.1, 1000),
+    arrays(np.float64, (3,), elements=floats(-10, 10, width=64), unique=True),
+)
+def test_get_set_minimal_bounding_ellipse_radius(a, b, center):
+    _test_get_set_minimal_bounding_sphere_radius(Ellipse(a, b, center))
+
+
+@given(
+    floats(0.1, 1000),
+    floats(0.1, 1000),
+    arrays(np.float64, (3,), elements=floats(-10, 10, width=64), unique=True),
+)
+def test_get_set_minimal_centered_bounding_ellipse_radius(a, b, center):
+    _test_get_set_minimal_bounding_sphere_radius(Ellipse(a, b, center), True)
 
 
 def test_inertia_tensor():