+++ tests/test_spheropolygon.py
@@ -8,7 +8,11 @@ from hypothesis.strategies import floats
 from pytest import approx
 from scipy.spatial import ConvexHull
 
-from conftest import EllipseSurfaceStrategy
+from conftest import (
+    EllipseSurfaceStrategy,
+    _test_get_set_minimal_bounding_sphere_radius,
+)
+from coxeter.families import RegularNGonFamily
 from coxeter.shapes import ConvexSpheropolygon
 
 
@@ -229,6 +233,43 @@ def test_perimeter_setter(unit_rounded_square):
         assert unit_rounded_square.radius == approx(1.0)
 
     testfun()
+
+
+@given(floats(0.1, 1000))
+def test_minimal_bounding_circle_regular_polygon(radius):
+    family = RegularNGonFamily()
+    for i in range(3, 10):
+        vertices = family.make_vertices(i)
+        rmax = np.max(np.linalg.norm(vertices, axis=-1)) + radius
+
+        poly = ConvexSpheropolygon(vertices, radius)
+        circle = poly.minimal_bounding_circle
+
+        assert np.isclose(rmax, circle.radius)
+        assert np.allclose(circle.center, 0)
+
+
+@given(floats(0.1, 1000))
+def test_minimal_centered_bounding_circle_regular_polygon(radius):
+    family = RegularNGonFamily()
+    for i in range(3, 10):
+        vertices = family.make_vertices(i)
+        rmax = np.max(np.linalg.norm(vertices, axis=-1)) + radius
+
+        poly = ConvexSpheropolygon(vertices, radius)
+        circle = poly.minimal_centered_bounding_circle
+
+        assert np.isclose(rmax, circle.radius)
+        assert np.allclose(circle.center, 0)
+
+
+@given(floats(0.1, 1000))
+def test_get_set_minimal_bounding_circle_radius(r):
+    family = RegularNGonFamily()
+    for i in range(3, 10):
+        _test_get_set_minimal_bounding_sphere_radius(
+            ConvexSpheropolygon(family.make_vertices(i), r)
+        )
 
 
 def test_inertia(unit_rounded_square):