+++ ChangeLog.rst
@@ -11,11 +11,20 @@ Added
 - Point-in-shape checks for circles.
 - Point-in-shape checks for ellipses.
 - Inertia tensors for 2D shapes that implement moments of inertia.
+- Add minimal bounding sphere for all shapes.
+- Add minimal centered bounding sphere calculations for all shapes except general polygons and polyhedra.
+- Enable getting and setting the circumsphere or bounding sphere radius of a polyhedron (for both types of bounding sphere).
 
 Changed
 ~~~~~~~
 
 - Ensure that hypothesis-based tests don't implicitly reuse pytest fixtures.
+
+Deprecated
+~~~~~~~
+
+- The circumsphere from center calculations (replaced by minimal centered bounding sphere).
+- The bounding_sphere property is deprecated in favor of minimal_bounding_sphere.
 
 
 v0.4.0 - 2020-10-14